refactor cli backend without transfer needs
