+++ src/aaz_dev/cli/model/atomic/__init__.py
@@ -4,4 +4,3 @@ from ._example import CLICommandExample
 from ._help import CLICommandHelp, CLICommandGroupHelp
 from ._profile import CLIAtomicProfile
 from ._resource import CLISpecsResource
-from ._module import CLIModule
