+++ src/aaz_dev/cli/model/atomic/_command.py
@@ -1,13 +1,14 @@
-from cli.model.atomic._fields import CLICommandNameField, CLIVersionField, CLICommandConfigurationField
-from cli.model.atomic._help import CLICommandHelp
-from cli.model.atomic._resource import CLISpecsResource
+from cli.model.common._fields import CLICommandNameField, CLIVersionField, CLICommandConfigurationField
 from schematics.models import Model
-from utils.stage import AAZStageField
 from schematics.types import ModelType, ListType, StringType, BooleanType
+from utils.stage import AAZStageField
+
+from ._help import CLICommandHelp
+from ._resource import CLISpecsResource
 
 
 class CLIAtomicCommandRegisterInfo(Model):
-    stage = AAZStageField(required=True)    # the stage used in code, usually it should be consist with command.stage
+    stage = AAZStageField(required=True)  # the stage used in code, usually it should be consist with command.stage
     confirmation = StringType(min_length=1)
 
     # TODO: add support for deprecate_info
@@ -18,7 +19,7 @@ class CLIAtomicCommandRegisterInfo(Model):
 
 class CLIAtomicCommand(Model):
     names = ListType(field=CLICommandNameField(), min_size=1, required=True)  # full name of a command
-    stage = AAZStageField()     # the stage of command
+    stage = AAZStageField()  # the stage of command
     help = ModelType(CLICommandHelp, required=True)
     register_info = ModelType(
         CLIAtomicCommandRegisterInfo,
