+++ src/aaz_dev/cli/model/view/__init__.py
@@ -1,0 +1,8 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+from ._module import CLIModule
+from ._profile import CLIViewProfile
+from ._command_group import CLIViewCommandGroup
+from ._command import CLIViewCommand
