+++ src/aaz_dev/cli/model/view/_command_group.py
@@ -1,0 +1,26 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+from cli.model.common._fields import CLICommandNameField
+from schematics.models import Model
+from schematics.types import ModelType, ListType, DictType
+from utils.stage import AAZStageField
+
+from ._command import CLIViewCommand
+
+
+class CLIViewCommandGroup(Model):
+    names = ListType(field=CLICommandNameField(), min_size=1, required=True)  # full name of a command group
+
+    command_groups = DictType(
+        field=ModelType("CLIViewCommandGroup"),
+        serialized_name="commandGroups",
+        deserialize_from="commandGroups"
+    )
+    commands = DictType(
+        field=ModelType(CLIViewCommand)
+    )
+
+    class Options:
+        serialize_when_none = False
